# api/api-swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express API with Swagger",
    "version": "1.0.0",
    "description": "REST API documentation using Swagger/OpenAPI"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server (HTTP)"
    },
    {
      "url": "https://localhost:3000",
      "description": "Development server (HTTPS)"
    }
  ],
  "paths": {
...
    "/api/cart": {
      "get": {
        "summary": "Get the cart for the current cart key",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XCartKeyHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "Cart contents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid cart key"
          }
        }
      },
      "delete": {
        "summary": "Clear all items in the current cart",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XCartKeyHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "Empty cart response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid cart key"
          }
        }
      }
    },
    "/api/cart/items": {
      "post": {
        "summary": "Add an item to the current cart",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XCartKeyHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "productId",
                  "quantity"
                ],
                "properties": {
                  "productId": {
                    "type": "integer"
                  },
                  "quantity": {
                    "type": "integer"
                  },
                  "cartKey": {
                    "type": "string",
                    "description": "Ignored if provided; header value is authoritative"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/api/cart/items/{productId}": {
      "delete": {
        "summary": "Remove an item from the current cart by product ID",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XCartKeyHeader"
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Product ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CartItem": {
        "type": "object",
        "required": [
          "cartItemId",
          "cartId",
          "productId",
          "quantity",
          "price"
        ],
        "properties": {
          "cartItemId": {
            "type": "integer",
            "description": "The unique identifier for the cart item"
          },
          "cartId": {
            "type": "integer",
            "description": "The ID of the cart containing the item"
          },
          "productId": {
            "type": "integer",
            "description": "The ID of the product in the cart"
          },
          "quantity": {
            "type": "integer",
            "description": "The quantity of the product in the cart"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "The price of one unit of the product"
          }
        }
      },
...
      "CartView": {
        "type": "object",
        "required": [
          "cartKey",
          "items",
          "itemCount",
          "subtotal"
        ],
        "properties": {
          "cartKey": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          },
          "itemCount": {
            "type": "integer"
          },
          "subtotal": {
            "type": "number",
            "format": "float"
          }
        }
      }
    },
    "parameters": {
      "XCartKeyHeader": {
        "in": "header",
        "name": "X-Cart-Key",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Opaque cart key used to identify a cart"
      }
    }
  },
...
}